Fix redirect to auth && add profile routers
Added profile route and settings child.
The token check + redirect to auth if logged out was redirecting the logged in user to home from any other route.
This has been fixed by changing the auth redirect logic, its no longer checking if isLoggedin is true, but if there is a token.